Remove email form, add linkedin link to contacts. Add same skills.
