+++ src/components/Contact/Contact.jsx
@@ -1,4 +1,3 @@
-import { useState } from 'react'
 import { Container, Row, Col } from 'react-bootstrap'
 import contactImg from '../../assets/img/contact-img.svg'
 import 'animate.css'
@@ -6,41 +5,6 @@ import TrackVisibility from 'react-on-screen'
 import './Contact.css'
 
 export const Contact = () => {
-    const formInitialDetails = {
-        name: '',
-        email: '',
-        message: '',
-    }
-    const [formDetails, setFormDetails] = useState(formInitialDetails)
-    const [buttonText, setButtonText] = useState('Send')
-    const [status, setStatus] = useState({})
-
-    const onFormUpdate = (category, value) => {
-        setFormDetails({
-            ...formDetails,
-            [category]: value,
-        })
-    }
-
-    const handleSubmit = async (e) => {
-        e.preventDefault()
-        setButtonText('Sending...')
-        let response = await fetch('http://localhost:5000/contact', {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json;charset=utf-8',
-            },
-            body: JSON.stringify(formDetails),
-        })
-        setButtonText('Send')
-        let result = await response.json()
-        setFormDetails(formInitialDetails)
-        if (result.code === 200) {
-            setStatus({ succes: true, message: 'Message sent successfully' })
-        } else {
-            setStatus({ succes: false, message: 'Something went wrong, please try again later.' })
-        }
-    }
 
     return (
         <section className="contact" id="contact">
@@ -62,32 +26,13 @@ export const Contact = () => {
                                     <div
                                         className={isVisible ? 'animate__animated animate__fadeIn' : 'animate__animated animate__fadeOut'}>
                                         <h2>Get In Touch</h2>
-                                        <form onSubmit={handleSubmit}>
-                                            <Row>
-                                                <Col size={12} sm={6} className="px-1">
-                                                    <input type="text" value={formDetails.name}
-                                                           placeholder="Name"
-                                                           onChange={(e) => onFormUpdate('name', e.target.value)} />
-                                                </Col>
-                                                <Col size={12} sm={6} className="px-1">
-                                                    <input type="email" value={formDetails.email}
-                                                           placeholder="Email Address"
-                                                           onChange={(e) => onFormUpdate('email', e.target.value)} />
-                                                </Col>
-                                                <Col size={12} className="px-1">
-                                                <textarea rows="6" value={formDetails.message} placeholder="Message"
-                                                          onChange={(e) => onFormUpdate('message', e.target.value)}></textarea>
-                                                    <button type="submit"><span>{buttonText}</span></button>
-                                                </Col>
-                                                {
-                                                    status.message &&
-                                                    <Col>
-                                                        <p className={status.success === false ? 'danger' : 'success'}>{status.message}</p>
-                                                    </Col>
-                                                }
-                                            </Row>
-                                        </form>
-                                    </div>}
+                                        <p className="stay-in-touch-text">Feel free to connect with me on <a
+                                            href="https://www.linkedin.com/in/andrii-vdovychenko" target="_blank"
+                                            rel="noreferrer">LinkedIn</a> to explore opportunities, share insights, and
+                                            network within the tech community.
+                                            Let's connect and stay in touch!</p>
+                                    </div>
+                                }
                             </TrackVisibility>
                         </Col>
                     </Row>
