+++ src/components/Contact/MailMe.jsx
@@ -1,0 +1,71 @@
+import { Col, Row } from 'react-bootstrap'
+import { useState } from 'react'
+
+// May be use it in the future?
+
+export const MailMe = () => {
+
+    const formInitialDetails = {
+        name: '',
+        email: '',
+        message: '',
+    }
+    const [formDetails, setFormDetails] = useState(formInitialDetails)
+    const [buttonText, setButtonText] = useState('Send')
+    const [status, setStatus] = useState({})
+
+    const onFormUpdate = (category, value) => {
+        setFormDetails({
+            ...formDetails,
+            [category]: value,
+        })
+    }
+
+    const handleSubmit = async (e) => {
+        e.preventDefault()
+        setButtonText('Sending...')
+        let response = await fetch('http://localhost:5000/contact', {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json;charset=utf-8',
+            },
+            body: JSON.stringify(formDetails),
+        })
+        setButtonText('Send')
+        let result = await response.json()
+        setFormDetails(formInitialDetails)
+        if (result.code === 200) {
+            setStatus({ succes: true, message: 'Message sent successfully' })
+        } else {
+            setStatus({ succes: false, message: 'Something went wrong, please try again later.' })
+        }
+    }
+
+    return (
+        <form onSubmit={handleSubmit}>
+            <Row>
+                <Col size={12} sm={6} className="px-1">
+                    <input type="text" value={formDetails.name}
+                           placeholder="Name"
+                           onChange={(e) => onFormUpdate('name', e.target.value)} />
+                </Col>
+                <Col size={12} sm={6} className="px-1">
+                    <input type="email" value={formDetails.email}
+                           placeholder="Email Address"
+                           onChange={(e) => onFormUpdate('email', e.target.value)} />
+                </Col>
+                <Col size={12} className="px-1">
+                                                <textarea rows="6" value={formDetails.message} placeholder="Message"
+                                                          onChange={(e) => onFormUpdate('message', e.target.value)}></textarea>
+                    <button type="submit"><span>{buttonText}</span></button>
+                </Col>
+                {
+                    status.message &&
+                    <Col>
+                        <p className={status.success === false ? 'danger' : 'success'}>{status.message}</p>
+                    </Col>
+                }
+            </Row>
+        </form>
+    )
+}
