+++ src/components/Skills/Skills.css
@@ -44,6 +44,15 @@
         }
     }
 
+    .react-multiple-carousel__arrow--left {
+        left: 0;
+        top: 20px;
+    }
+    .react-multiple-carousel__arrow--right {
+        right: 0;
+        top: 20px;
+    }
+
     .background-image-left {
         top: 28%;
         position: absolute;
